move shortened augment into main
